refactor(control-ui): split cron run-history rendering into view-runs
